Finished turret prefab, upgraded project and cleaned code

+++ Unity/ProjectConstantine/Assets/Scripts/Enemies/EnemyProjectileBaseData.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 public class EnemyProjectileBaseData : ScriptableObject
 {
     public float MovementSpeed = 5f;
-    public float ProjectileDamage = 10f;
+    public int ProjectileDamage = 10;
     public float ProjectileRange = 10;
     public GameObject ProjectilePrefab;
 }
